# doc/openapi/openapi-iam-v1.yaml
openapi: 3.0.3
info:
  title: IAM System API
  description: |
    Generic Multi-Tenant Identity and Access Management System API.

    ## Overview
    This API provides authentication, authorization, and user management capabilities
    for multi-tenant applications.

    ## Authentication
    Most endpoints require a JWT Bearer token in the Authorization header.
    Tokens are obtained via the OTP-based login flow (`/login` → `/login/{id}/verify-otp`).

    ## Multi-Tenancy
    The `X-Tenant-Code` header is required for authentication endpoints to identify
    the tenant context. For authenticated requests, the tenant is extracted from the JWT.

    ## Rate Limiting
    All endpoints are rate-limited. See the rate limit headers in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: Proprietary
    url: https://example.com/license
servers:
  - url: https://api.example.com/api/iam/v1
    description: Production
  - url: https://api-staging.example.com/api/iam/v1
    description: Staging
  - url: http://localhost:8080/api/iam/v1
    description: Development
...
paths:
...
  /auth/logout-all:
    post:
      tags: [Authentication]
      summary: Logout all sessions
      description: |
        Revokes **all** sessions for the authenticated user by:
        - Revoking all refresh tokens in the database
        - Adding user to user-level blacklist in Redis (invalidates all access tokens)
        - Invalidating all user sessions

        **Use case**: Security incident response, password change, or user-initiated
        "logout from all devices" action.

        **Idempotent**: Returns 200 even if the user has no active sessions.

        **Security**: Requires valid JWT in Authorization header. User ID is
        automatically extracted from the JWT claims.
      operationId: logoutAll
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All sessions logged out successfully (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "All sessions logged out successfully"
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          nullable: true
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: object
          properties:
            code:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/Meta'
...
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required or failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'